+++ src/components/Button/ButtonAligner.tsx
@@ -1,9 +1,16 @@
 import React from "react";
-import Aligner, { type AlignerProps } from "components/_private/Aligner";
+import Aligner from "components/_private/Aligner";
+import type * as T from "./Button.types";
 import s from "./Button.module.css";
 
-const ButtonAligner = (props: Omit<AlignerProps, "side"> & { position?: AlignerProps["side"] }) => {
-	return <Aligner {...props} side={props.position} className={[s.aligner, props.className]} />;
+const ButtonAligner = (props: T.AlignerProps) => {
+	return (
+		<Aligner
+			{...props}
+			side={props.side || props.position}
+			className={[s.aligner, props.className]}
+		/>
+	);
 };
 
 export default ButtonAligner;