+++ src/components/MenuItem/MenuItemAligner.tsx
@@ -2,8 +2,10 @@ import React from "react";
 import Aligner, { type AlignerProps } from "components/_private/Aligner";
 import s from "./MenuItem.module.css";
 
-const MenuItemAligner = (props: Omit<AlignerProps, "side">) => {
-	return <Aligner {...props} side="inline" className={[s.aligner, props.className]} />;
+const MenuItemAligner = (props: AlignerProps) => {
+	return (
+		<Aligner {...props} side={props.side || "inline"} className={[s.aligner, props.className]} />
+	);
 };
 
 export default MenuItemAligner;